requirements.txt
pydantic==1.8.2
PyYAML==6.0
elasticsearch==7.15.2
psycopg2-binary==2.9.1